matroskademux: Use segment.offset instead of segment.start
Set segment.offset to the PTS of the first frame found in the file.
(Previously segment.start was set to this instead). Set segment.start
and segment.time always to zero.

In the case of a cropped Matroska file -- let's say we had a 5 minutes
long video where we cropped the second minute (as could be done for
streaming purposes), this way the file can still be played instantly if
opened with a GStreamer player (the first frame PTS has running_time =
0) and stream times are now correct (stream time of buffer PTS=60s is
now 60s, previously it was zero).