qtdemux: Add MSE-style flush
The abort() method of SourceBuffer in Media Source Extensions is
expected to flush the demuxer and discard the current fragment,
if any. The configuration of tracks, if any, should be preserved.

Before this patch, qtdemux was only to able to handle correctly a flush
when it was a direct consequence of a byte seek that qtdemux itself sent
upstream in response to a time seek.

This patch modifies the handling of flush events in such a way that
qtdemux can differentiate between these flushes that are a consequence
of a seek and flushes coming from the application on its own (like in
MSE).

In the latter case, all unparsed buffered data is discarded, the sample
table is resetted and the demuxer prepares to read a new atom from the
beginning.

https://bugzilla.gnome.org/show_bug.cgi?id=795424