qtdemux: handle edits with unspecified end in gst_qtdemux_map_and_push_segments
MP4 allows the duration field of an edit (segment_duration) to be set to
zero. In that case it's considered that the edit finishes where the
track ends.

Quoting the spec (ISO/IEC 14496 Part 12, Fifth edition, 2015):

A non‐empty edit may insert a portion of the media timeline that is not
present in the initial movie, and is present only in subsequent movie
fragments. Particularly in an empty initial movie of a fragmented movie
file (when there are no media samples yet present), the segment_duration
of this edit may be zero, whereupon the edit provides the offset from
media composition time to movie presentation time, for the movie and
subsequent movie fragments. It is recommended that such an edit be used
to establish a presentation time of 0 for the first presented sample,
when composition offsets are used.