qtdemux tests: Add edit list test battery
Edit list handling has enough complexity to warrant some tests.

This patch introduces a test battery that generates a variety of edit
lists in different types of files (fragmented or not fragmented) and
tests them in both push and pull mode.

Since the valid combinations of all the test variables are quite a few
(35 tests), code generation is used for the actual test functions.
The code generator also defines the expected failures due to bugs not
currently fixed.

https://bugzilla.gnome.org/show_bug.cgi?id=794902