qtdemux: Update the segment for KEY_UNIT seeks via gst_segment_do_seek()
Manually we were not adjusting the segment.offset accordingly, which
caused jumps in running time if cur_type==GST_SEEK_TYPE_NONE in the
seek.

The demuxer segment is now always the actual seek segment instead of any
manually adjusted segment, and it is taken as the base for generating
the per-stream / per edit list segment segments.

https://bugzilla.gnome.org/show_bug.cgi?id=796559