qtdemux: Calculate per-stream segment base directly from the segment
There is no need to manually calculate it, we have the expected value
directly available in the demuxer segment. And the previous calculations
were not adjusting the segment.offset accordingly, thus shifting the
running time.

As a nice side-effect this makes it unnecessary to accumulate the base
time of the different edit list segments.

https://bugzilla.gnome.org/show_bug.cgi?id=796559